v0.0.7.0

+++ app/src/main/java/com/example/testcdc/dao/SignalInfoDao.java
@@ -43,6 +43,6 @@ public interface SignalInfoDao {
     }
 
     @Query("DELETE FROM signal_info WHERE cid = :cid")
-    void deleteBycid(int cid);
+    void deleteBycid(long cid);
 
 }
